backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-dotenv==1.0.1
python-multipart==0.0.12
google-generativeai==0.8.3
aiofiles==24.1.0
pydantic==2.9.2
openai-whisper
fpdf2
python-docx
PyPDF2
